+++ lib/main.dart
@@ -9,6 +9,8 @@
 
 import 'package:flutter/material.dart';
 import 'package:music_player/screens/category_selector.dart';
+import 'package:music_player/services/audio/audio_player_service.dart';
+import 'package:music_player/services/audio/just_audio_player.dart';
 import 'package:music_player/services/playlists/hardcoded_playlists_service.dart';
 import 'package:music_player/services/playlists/playlists_service.dart';
 import 'package:provider/provider.dart';
@@ -18,13 +20,22 @@ void main() => runApp(MyApp());
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      home: Provider<PlaylistsService>(
-          create: (_) {
-            return HardcodedPlaylistsService();
+    return MultiProvider(
+      providers: [
+        Provider<PlaylistsService>(
+          create: (_) => HardcodedPlaylistsService(),
+        ),
+        Provider<AudioPlayerService>(
+          create: (_) => JustAudioPlayer(),
+          dispose: (_, value) {
+            (value as JustAudioPlayer).dispose();
           },
-          child: CategorySelector()),
+        ),
+      ],
+      child: MaterialApp(
+        debugShowCheckedModeBanner: false,
+        home: CategorySelector(),
+      ),
     );
   }
 }