+++ lib/screens/category_selector.dart
@@ -8,9 +8,8 @@
  */
 
 import 'package:flutter/material.dart';
-import 'package:just_audio/just_audio.dart';
-import 'package:music_player/screens/commons/player_buttons.dart';
-import 'package:music_player/screens/player.dart';
+import 'package:music_player/screens/commons/player_buttons_container.dart';
+import 'package:music_player/screens/playlist_screen.dart';
 import 'package:music_player/services/playlists/playlists_service.dart';
 import 'package:provider/provider.dart';
 
@@ -18,38 +17,20 @@ import 'package:provider/provider.dart';
 ///
 /// Current categories are:
 ///  - all items;
-class CategorySelector extends StatefulWidget {
-  @override
-  _CategorySelectorState createState() => _CategorySelectorState();
-}
-
-class _CategorySelectorState extends State<CategorySelector> {
-  late AudioPlayer _audioPlayer;
-
-  @override
-  void initState() {
-    super.initState();
-    _audioPlayer = AudioPlayer();
-  }
-
-  @override
-  void dispose() {
-    _audioPlayer.dispose();
-    super.dispose();
-  }
-
+class CategorySelector extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(),
       body: Center(
         child: SafeArea(
-          child: Consumer<PlaylistsService>(
-            builder: (__, value, _) {
-              return Column(
-                children: [
-                  Expanded(
-                    child: ListView(
+          child: PlayerButtonsContainer(
+            child: Consumer<PlaylistsService>(
+              builder: (__, value, _) {
+                return Column(
+                  children: [
+                    ListView(
+                      shrinkWrap: true,
                       children: [
                         ListTile(
                           title: Text("All items"),
@@ -57,26 +38,31 @@ class _CategorySelectorState extends State<CategorySelector> {
                             Navigator.of(context).push(
                               MaterialPageRoute(
                                 builder: (context) =>
-                                    Player(_audioPlayer, value.allItems),
+                                    PlaylistScreen(value.allItems),
                               ),
                             );
                           },
                         ),
-                      ],
+                      ]..addAll(
+                          value.playlists.keys.map((playlistName) {
+                            return ListTile(
+                              title: Text(playlistName),
+                              onTap: () {
+                                Navigator.of(context).push(
+                                  MaterialPageRoute(
+                                    builder: (context) => PlaylistScreen(
+                                        value.playlists[playlistName]!),
+                                  ),
+                                );
+                              },
+                            );
+                          }),
+                        ),
                     ),
-                  ),
-                  StreamBuilder<bool>(
-                      stream: _audioPlayer.playingStream,
-                      builder: (context, snapshot) {
-                        // If we are not playing, do not show the player buttons
-                        if (snapshot.hasData && (snapshot.data ?? false))
-                          return PlayerButtons(_audioPlayer);
-                        else
-                          return Container();
-                      }),
-                ],
-              );
-            },
+                  ],
+                );
+              },
+            ),
           ),
         ),
       ),