+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.5]
+### Changed
+- Refactor the player behind an interface
+
+### Fixed
+- User could not play a different playlist if one is loaded already
+
+## [0.0.4] - add Provider and move hardcoded data
+### Added
+- Provider
+
+### Changed
+- moved hardcoded data to its own file
+
 ## [0.0.3] - migrate to sound null safety
 ### Added
 - license
@@ -20,5 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - a row of buttons to play, skip, shuffle, and loop.
 
+[0.0.4]: https://github.com/mvolpato/the-player/releases/tag/0.0.4
+[0.0.3]: https://github.com/mvolpato/the-player/releases/tag/0.0.3
 [0.0.2]: https://github.com/mvolpato/the-player/releases/tag/0.0.2
 [0.0.1]: https://github.com/mvolpato/the-player/releases/tag/0.0.1